--- src/connector/signal.ts
@@ -1,13 +1,0 @@
-import { Trickle } from './client';
-export { Trickle };
-
-export interface Signal {
-  onnegotiate?: (jsep: RTCSessionDescriptionInit) => void;
-  ontrickle?: (trickle: Trickle) => void;
-
-  join(sid: string, uid: null | string): Promise<any>;
-  offer(offer: RTCSessionDescriptionInit): Promise<RTCSessionDescriptionInit>;
-  answer(answer: RTCSessionDescriptionInit): void;
-  trickle(trickle: Trickle): void;
-  close(): void;
-}
